modify some of comment-module according reviews some serialization problem arise
